# A small performance improvement and more sensible benchmarks