Let's exercise only arguments that are in range...
```
Run on (4 X 3310 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
BM_Elbdj                 274 ns        274 ns    9976000
BM_Elbdj                 275 ns        274 ns    9976000
BM_Elbdj                 273 ns        271 ns    9976000
BM_Elbdj                 272 ns        274 ns    9976000
BM_Elbdj                 273 ns        271 ns    9976000
BM_Elbdj                 273 ns        272 ns    9976000
BM_Elbdj                 272 ns        272 ns    9976000
BM_Elbdj                 272 ns        272 ns    9976000
BM_Elbdj                 274 ns        272 ns    9976000
BM_Elbdj                 272 ns        272 ns    9976000
BM_Elbdj_mean            273 ns        272 ns         10
BM_Elbdj_median          273 ns        272 ns         10
BM_Elbdj_stddev            1 ns          1 ns         10
```